Release 2.1.0-rc.8.1 :rocket:
- Milestone Details: https://github.com/mean-expert-official/loopback-sdk-builder/milestone/31?closed=1

- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/306
- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/304
- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/303
- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/302
