+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file is created to keep history of the LoopBack SDK Builder, it does not consider or keeps any history of its parent module `loopback-sdk-angular`.
 
+## Release 2.1.0-rc.8.1
+
+- Milestone Details: https://github.com/mean-expert-official/loopback-sdk-builder/milestone/31?closed=1
+
+- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/306
+- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/304
+- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/303
+- Fix: https://github.com/mean-expert-official/loopback-sdk-builder/issues/302
+
 ## Release 2.1.0-rc.8
 
 - Milestone Details: https://github.com/mean-expert-official/loopback-sdk-builder/milestone/30?closed=1
