+++ lib/angular2/shared/models/base.ejs
@@ -19,7 +19,7 @@ export interface AccessTokenInterface {
     ttl?: number;
     issuedAt?: any;
     created: any;
-    userId?: string;
+    userId?: number;
     rememberMe?: boolean;
 }
 
@@ -28,7 +28,7 @@ export class AccessToken implements AccessTokenInterface {
   ttl: number = 1209600;
   created: Date = new Date(0);
   userId: number = 0;
-  user: User = null;
+  user: any = null;
   constructor(data?: AccessTokenInterface) {
     Object.assign(this, data);
   }
