+++ tests/fireloop/server/model-config.json
@@ -21,7 +21,7 @@
   },
   "AccessToken": {
     "dataSource": "db",
-    "public": true
+    "public": false
   },
   "ACL": {
     "dataSource": "db",
