+++ tests/ng2web/src/app/shared/sdk/services/core/auth.service.ts
@@ -31,8 +31,8 @@ export class LoopBackAuth {
     this.token.rememberMe = value;
   }
 
-  public getToken(): AccessToken {
-    return <AccessToken> this.token;
+  public getToken(): SDKToken {
+    return <SDKToken> this.token;
   }
 
   public getAccessTokenId(): any {
