+++ tests/ng2web/src/app/shared/sdk/services/core/real.time.ts
@@ -40,7 +40,7 @@ export class RealTime {
   * @description
   * Will trigger when Real-Time Service is disconnected from server.
   **/
-  onDisconnect(): Observable<null> {
+  onDisconnect(): Observable<any> {
     return this.connection.sharedObservables.sharedOnDisconnect;
   }
   /**
@@ -48,7 +48,7 @@ export class RealTime {
   * @description
   * Will trigger when Real-Time Service is not authorized from server.
   **/
-  onAuthenticated(): Observable<null> {
+  onAuthenticated(): Observable<any> {
     return this.connection.sharedObservables.sharedOnAuthenticated;
   }
   /**
@@ -56,7 +56,7 @@ export class RealTime {
   * @description
   * Will trigger when Real-Time Service is not authorized from server.
   **/
-  onUnAuthorized(): Observable<null> {
+  onUnAuthorized(): Observable<any> {
     return this.connection.sharedObservables.sharedOnUnAuthorized;
   }
   /**
