--- tests/ng2web/src/app/shared/sdk/services/custom/AccessToken.ts
@@ -1,72 +1,0 @@
-/* tslint:disable */
-import { Injectable, Inject, Optional } from '@angular/core';
-import { Http, Response } from '@angular/http';
-import { SDKModels } from './SDKModels';
-import { BaseLoopBackApi } from '../core/base.service';
-import { LoopBackConfig } from '../../lb.config';
-import { LoopBackAuth } from '../core/auth.service';
-import { LoopBackFilter,  } from '../../models/BaseModels';
-import { JSONSearchParams } from '../core/search.params';
-import { ErrorHandler } from '../core/error.service';
-import { Subject } from 'rxjs/Subject';
-import { Observable } from 'rxjs/Rx';
-import { AccessToken } from '../../models/AccessToken';
-import { SocketConnection } from '../../sockets/socket.connections';
-import { User } from '../../models/User';
-
-
-/**
- * Api services for the `AccessToken` model.
- */
-@Injectable()
-export class AccessTokenApi extends BaseLoopBackApi {
-
-  constructor(
-    @Inject(Http) protected http: Http,
-    @Inject(SocketConnection) protected connection: SocketConnection,
-    @Inject(SDKModels) protected models: SDKModels,
-    @Inject(LoopBackAuth) protected auth: LoopBackAuth,
-    @Inject(JSONSearchParams) protected searchParams: JSONSearchParams,
-    @Optional() @Inject(ErrorHandler) protected errorHandler: ErrorHandler
-  ) {
-    super(http,  connection,  models, auth, searchParams, errorHandler);
-  }
-
-  /**
-   * Fetches belongsTo relation user.
-   *
-   * @param any id PersistedModel id
-   *
-   * @param boolean refresh 
-   *
-   * @returns object An empty reference that will be
-   *   populated with the actual data once the response is returned
-   *   from the server.
-   *
-   * <em>
-   * (The remote method definition does not provide any description.
-   * This usually means the response is a `AccessToken` object.)
-   * </em>
-   */
-  public getUser(id: any, refresh: any = {}): Observable<any> {
-    let _method: string = "GET";
-    let _url: string = LoopBackConfig.getPath() + "/" + LoopBackConfig.getApiVersion() +
-    "/AccessTokens/:id/user";
-    let _routeParams: any = {
-      id: id
-    };
-    let _postBody: any = {};
-    let _urlParams: any = {};
-    if (refresh) _urlParams.refresh = refresh;
-    let result = this.request(_method, _url, _routeParams, _urlParams, _postBody);
-    return result;
-  }
-
-  /**
-   * The name of the model represented by this $resource,
-   * i.e. `AccessToken`.
-   */
-  public getModelName() {
-    return "AccessToken";
-  }
-}
